requirements.txt
Flask==2.0.1
python-docx==0.8.11
fpdf==1.7.2
